// src/app/Interfaces/alimentos.interface.ts
export interface AlimentosIngrediente {
    idMeal: string;
    strMeal: string;
    strMealThumb: string;
}
export interface IdAlimento {
    idMeal: string;
}
export interface DetalleAlimento {
    dateModified: any;
    idMeal: string;
    strArea: string;
    strCategory: string;
    strCreativeCommonsConfirmed: any;
    strDrinkAlternate: any;
    strImageSource: any;
    strIngredient1: string;
    strIngredient2: string;
    strIngredient3: string;
    strIngredient4: string;
    strIngredient5: string;
    strIngredient6: string;
    strIngredient7: string;
    strIngredient8: string;
    strIngredient9: string;
    strIngredient10: string;
    strIngredient11: string;
    strIngredient12: string;
    strIngredient13: string;
    strIngredient14: string;
    strIngredient15: string;
    strIngredient16: any;
    strIngredient17: any;
    strIngredient18: any;
    strIngredient19: any;
    strIngredient20: any;
    strInstructions: string;
    strMeal: string;
    strMealThumb: string;
    strMeasure1: string;
    strMeasure2: string;
    strMeasure3: string;
    strMeasure4: string;
    strMeasure5: string;
    strMeasure6: string;
    strMeasure7: string;
    strMeasure8: string;
    strMeasure9: string;
    strMeasure10: string;
    strMeasure11: string;
    strMeasure12: string;
    strMeasure13: string;
    strMeasure14: string;
    strMeasure15: string;
    strMeasure16: any;
    strMeasure17: any;
    strMeasure18: any;
    strMeasure19: any;
    strMeasure20: any;
    strSource: any;
    strTags: string;
    strYoutube: string;
}